requirements.txt
pandas
scikit-learn
xgboost
seaborn
mlflow
flask
torch
transformers
Pillow

-e .
